MemorySegment object pooling
See https://issues.apache.org/jira/browse/FLINK-3322

Also note that the problem was much worse now, because the
taskmanager.memory.preallocate option is gone, and somehow the overhead
also seemed to be bigger now.

This adds a per-owner pooling. Per-MemoryMAnager would be better, but
it's fine for the paper. The only issue now is that if an operator needs
less memory in a later iteration step then before, others still cannot
take its memory, because it's still stuck in its pool. Note that this
would not have any impact in vanilla Flink, but we have a further change
on this branch where the solution set can grab more memory than its
share, and it would steal this from other operators' free memory, which
might be less because of pooling per owner.