Change memory management to allow the solution set to grab more memory than its default budget
Before this change, Flink divided memory equally among all operators.
However, not all operators need all this memory. The not needed memory
can be given to the solution set, since if the solution set memory runs
out we would crash anyhow.

We also need to change other operators to not grab all their budget from
the MemoryManager eagerly. So far, I changed TempBarrier, which might be
enough for the paper.